refactor: more like mvc
